test(BA-6979): move the replica-group history tests into TestSchedulingHistoryRepository
They exercise SchedulingHistoryRepository, so they belong in that class
next to the session, deployment and route cases rather than in a file of
their own.
